# lib/api-zod/src/generated/types/aeoData.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AuditNexus API — SEO/GEO/AEO audit pipeline
 * OpenAPI spec version: 0.1.0
 */

export interface AeoData {
  /** @nullable */
  featuredSnippetScore?: number | null;
  /** @nullable */
  structuredDataScore?: number | null;
  /** @nullable */
  faqSchemaScore?: number | null;
  /** @nullable */
  voiceSearchScore?: number | null;
  issues?: string[];
  recommendations?: string[];
}