# lib/api-zod/src/generated/types/audit.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AuditNexus API — SEO/GEO/AEO audit pipeline
 * OpenAPI spec version: 0.1.0
 */
import type { AeoData } from './aeoData';
import type { AuditStatus } from './auditStatus';
import type { GeoData } from './geoData';
import type { SeoData } from './seoData';

export interface Audit {
  id: number;
  url: string;
  /** @nullable */
  clientName?: string | null;
  /** @nullable */
  clientEmail?: string | null;
  status: AuditStatus;
  /** @nullable */
  seoScore?: number | null;
  /** @nullable */
  geoScore?: number | null;
  /** @nullable */
  aeoScore?: number | null;
  /** @nullable */
  overallScore?: number | null;
  seoData?: SeoData;
  geoData?: GeoData;
  aeoData?: AeoData;
  createdAt: string;
  updatedAt: string;
}