# lib/api-zod/src/generated/types/auditInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AuditNexus API — SEO/GEO/AEO audit pipeline
 * OpenAPI spec version: 0.1.0
 */

export interface AuditInput {
  /** @minLength 1 */
  url: string;
  clientName?: string;
  clientEmail?: string;
}