# lib/api-zod/src/generated/types/clientInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AuditNexus API — SEO/GEO/AEO audit pipeline
 * OpenAPI spec version: 0.1.0
 */

export interface ClientInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  email: string;
  company?: string;
  phone?: string;
  website?: string;
  notes?: string;
}