# lib/api-zod/src/generated/types/contractUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AuditNexus API — SEO/GEO/AEO audit pipeline
 * OpenAPI spec version: 0.1.0
 */
import type { ContractUpdateStatus } from './contractUpdateStatus';

export interface ContractUpdate {
  status?: ContractUpdateStatus;
  signedAt?: string;
  content?: string;
  startDate?: string;
  endDate?: string;
}