# lib/api-zod/src/generated/types/geoData.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AuditNexus API — SEO/GEO/AEO audit pipeline
 * OpenAPI spec version: 0.1.0
 */

export interface GeoData {
  /** @nullable */
  localPresenceScore?: number | null;
  /** @nullable */
  citationsScore?: number | null;
  /** @nullable */
  napConsistencyScore?: number | null;
  /** @nullable */
  localKeywordsScore?: number | null;
  issues?: string[];
  recommendations?: string[];
}