# lib/api-zod/src/generated/types/proposalInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AuditNexus API — SEO/GEO/AEO audit pipeline
 * OpenAPI spec version: 0.1.0
 */
import type { ProposalService } from './proposalService';

export interface ProposalInput {
  auditId?: number;
  clientId?: number;
  /** @minLength 1 */
  clientName: string;
  /** @minLength 1 */
  clientEmail: string;
  /** @minLength 1 */
  title: string;
  description?: string;
  services?: ProposalService[];
  totalPrice: number;
  currency?: string;
  validUntil?: string;
}