# lib/api-zod/src/generated/types/proposalUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AuditNexus API — SEO/GEO/AEO audit pipeline
 * OpenAPI spec version: 0.1.0
 */
import type { ProposalService } from './proposalService';
import type { ProposalUpdateStatus } from './proposalUpdateStatus';

export interface ProposalUpdate {
  title?: string;
  description?: string;
  services?: ProposalService[];
  totalPrice?: number;
  status?: ProposalUpdateStatus;
  validUntil?: string;
}